split resolver type replay
