+++ src/typechecker/declaration_collection_resolver_tasks/replay_kinds.rs
@@ -2,72 +2,6 @@ use super::*;
 
 impl TypeChecker {
     #[cfg(test)]
-    pub(in crate::typechecker) fn collect_resolver_type_declaration_metadata_tasks(
-        decls: &[Declaration],
-    ) -> Vec<ResolverTypeDeclarationMetadataTask<'_>> {
-        let mut tasks = Vec::new();
-        for decl in decls {
-            Self::push_resolver_type_metadata_task(decl, &mut tasks);
-        }
-        tasks
-    }
-
-    pub(in crate::typechecker) fn push_resolver_type_replay_tasks<'a>(
-        decl: &'a Declaration,
-        type_tasks: &mut Vec<ResolverTypeDeclarationMetadataTask<'a>>,
-        type_reference_tasks: &mut Vec<ResolverTypeReferenceValidationTask<'a>>,
-    ) -> bool {
-        match decl {
-            Declaration::Struct {
-                name, fields, span, ..
-            } => {
-                type_tasks.push(ResolverTypeDeclarationMetadataTask::Struct {
-                    name,
-                    fields,
-                    span: *span,
-                });
-                type_reference_tasks.push(ResolverTypeReferenceValidationTask::Struct {
-                    name,
-                    fields,
-                    span: *span,
-                });
-                true
-            }
-            Declaration::Enum { name, span, .. } => {
-                type_tasks.push(ResolverTypeDeclarationMetadataTask::Enum { name, span: *span });
-                type_reference_tasks
-                    .push(ResolverTypeReferenceValidationTask::Enum { name, span: *span });
-                true
-            }
-            _ => false,
-        }
-    }
-
-    #[cfg(test)]
-    fn push_resolver_type_metadata_task<'a>(
-        decl: &'a Declaration,
-        tasks: &mut Vec<ResolverTypeDeclarationMetadataTask<'a>>,
-    ) -> bool {
-        match decl {
-            Declaration::Struct {
-                name, fields, span, ..
-            } => {
-                tasks.push(ResolverTypeDeclarationMetadataTask::Struct {
-                    name,
-                    fields,
-                    span: *span,
-                });
-                true
-            }
-            Declaration::Enum { name, span, .. } => {
-                tasks.push(ResolverTypeDeclarationMetadataTask::Enum { name, span: *span });
-                true
-            }
-            _ => false,
-        }
-    }
-
-    #[cfg(test)]
     pub(in crate::typechecker) fn collect_resolver_behavior_declaration_metadata_tasks(
         decls: &[Declaration],
     ) -> Vec<ResolverBehaviorDeclarationMetadataTask<'_>> {
